TypeScript Phase 16: Types file organization (#387)
