+++ packages/fastify-vite/src/index.test-d.ts
