+++ packages/fastify-vite/src/types/bundle.ts
@@ -1,0 +1,14 @@
+import type { Manifest } from 'vite'
+import type { ExtendedResolvedViteConfig } from './vite-configs.ts'
+
+export interface Bundle {
+  manifest?: Manifest
+  indexHtml?: string
+  dir?: string
+}
+
+export type BundleConfig = {
+  dev: boolean
+  vite: ExtendedResolvedViteConfig
+  root: string
+}
