+++ packages/fastify-vite/src/types/html.ts
@@ -1,0 +1,3 @@
+import type { ReplyDotRenderResult } from './reply.ts'
+
+export type HtmlTemplateFunction = (data?: ReplyDotRenderResult) => string
